Wait for the git path to appear on startup.
Without this, the bind companion container will simply crash until
git-sync has finished doing its job. There is a time limit (default: 5
minutes) to ensure we don't wait indefinitely without any kind of
warning.

This change means that the time in which the container is not ready to
server queries yet is extended, which emphasizes the need for a health
check handler.