Separate watch and build paths.
inotify doesn't seem to notice when a symlink is updated, probably
because the fsnotify API instrucs it to follow said symlink. Since the
symlink destination is not changed, we never notice any DNS file
updates.

Watching the directory above should work, but requires to keep track of
the path to the Makefile separately.